changed value from byte to int

+++ ReadASCIIString/ReadASCIIString.ino
@@ -46,15 +46,7 @@ int i = 1;
 // initializes pickup position
 int PickupPosition = 50;
 
-// initializes char array with a size of 150
-// int RPMmap[150] = {};
-// int ADVmap[150] = {};
-// int SAVE_TO_EEPROM_ODD[150] = {};
-// int SAVE_TO_EEPROM_EVEN[150] = {};
-// int TO_EEPROM[150] = {};
-
-
-byte value;
+int value;
 
 void setup() {
   // initialize serial:
@@ -71,17 +63,6 @@ void setup() {
   Serial.println(""); 
   Serial.println("Reading values from USB---------. Enter: rpm , advance");
 }
-
-// void readEEPROM() {
-//     for(int i = 0; i <= 512; ++i){
-//         value = EEPROM.read(i);
-//         Serial.print(i);
-//         Serial.print("\t");
-//         Serial.print(value, DEC);
-//         Serial.println();
-//     }
-
-// }
 
 void loop() {
   // if there's any serial available, read it:
